+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 - Reject invalid queue size exporterhelper (#4799)
 - Transform configmapprovider.Retrieved interface to a struct (#4789)
 
+## 🚩 Deprecations 🚩
+
+- Deprecate `pdata.NumberDataPoint.Type()` and `pdata.Exemplar.Type()` in favor of `NumberDataPoint.ValueType()` and
+  `Exemplar.ValueType()` (#4850)
+
 ## v0.44.0 Beta
 
 ## 🛑 Breaking changes 🛑